+++ src/main/java/org/recordy/server/common/config/WebConfig.java
@@ -1,7 +1,7 @@
 package org.recordy.server.common.config;
 
 import lombok.RequiredArgsConstructor;
-import org.recordy.server.auth.security.AccessTokenArgumentResolver;
+import org.recordy.server.auth.security.resolver.AccessTokenArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
