+++ src/main/java/org/recordy/server/record/controller/RecordController.java
@@ -2,7 +2,7 @@ package org.recordy.server.record.controller;
 
 
 import lombok.RequiredArgsConstructor;
-import org.recordy.server.auth.security.UserId;
+import org.recordy.server.auth.security.resolver.UserId;
 import org.recordy.server.record.controller.dto.request.RecordCreateRequest;
 import org.recordy.server.record.domain.File;
 import org.recordy.server.record.domain.Record;
