+++ src/test/java/org/recordy/server/auth/security/handler/AuthFilterExceptionHandlerTest.java
@@ -1,6 +1,5 @@
 package org.recordy.server.auth.security.handler;
 
-import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.recordy.server.auth.exception.AuthException;
@@ -8,10 +7,7 @@ import org.recordy.server.common.message.ErrorMessage;
 import org.recordy.server.mock.FakeContainer;
 import org.springframework.mock.web.MockHttpServletResponse;
 
-import java.io.IOException;
-
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class AuthFilterExceptionHandlerTest {
